back/requirements.txt
asgiref==3.2.10
Django==3.0.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
